requirements.txt
# Podstawy konfiguracji i środowiska
python-dotenv
requests

# Microsoft Graph
msgraph-core==0.2.2

# Uwierzytelnianie i SDK Azure
azure-identity
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-costmanagement

# gRPC + Protobuf (dobrane wersje działające na Win + Python 3.11)
grpcio==1.62.2
grpcio-tools==1.62.2
protobuf==4.25.3
